src/requirements.txt
psycopg2-binary
requests
git+https://github.com/Crypto-toolbox/bitex@dev
python-kucoin
lxml